fix(container): breakdown files to avoid importing `vite`
