+++ .changeset/happy-boxes-collect.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Fixes an issue with the container APIs where a runtime error was thrown during the build, when using `pnpm` as package manager.
