+++ packages/astro/src/container/index.ts
@@ -14,7 +14,7 @@ import type {
 	SSRManifest,
 	SSRResult,
 } from '../@types/astro.js';
-import { validateConfig } from '../core/config/config.js';
+import { validateConfig } from '../core/config/validate.js';
 import { ASTRO_CONFIG_DEFAULTS } from '../core/config/schema.js';
 import { Logger } from '../core/logger/core.js';
 import { nodeLogDestination } from '../core/logger/node.js';
@@ -208,7 +208,7 @@ type AstroContainerConstructor = {
 	renderers?: SSRLoadedRenderer[];
 	manifest?: AstroContainerManifest;
 	resolve?: SSRResult['resolve'];
-	astroConfig: AstroConfig;
+	astroConfig?: AstroConfig;
 };
 
 export class experimental_AstroContainer {
@@ -253,10 +253,10 @@ export class experimental_AstroContainer {
 		});
 	}
 
-	async #containerResolve(specifier: string, astroConfig: AstroConfig): Promise<string> {
+	async #containerResolve(specifier: string, astroConfig?: AstroConfig): Promise<string> {
 		const found = this.#pipeline.manifest.entryModules[specifier];
 		if (found) {
-			return new URL(found, astroConfig.build.client).toString();
+			return new URL(found, astroConfig?.build.client).toString();
 		}
 		return found;
 	}
@@ -329,10 +329,10 @@ export class experimental_AstroContainer {
 	private static async createFromManifest(
 		manifest: SSRManifest
 	): Promise<experimental_AstroContainer> {
-		const astroConfig = await validateConfig(ASTRO_CONFIG_DEFAULTS, process.cwd(), 'container');
+		// const astroConfig = await validateConfig(ASTRO_CONFIG_DEFAULTS, process.cwd(), 'container');
 		const container = new experimental_AstroContainer({
 			manifest,
-			astroConfig,
+			// astroConfig,
 		});
 		container.#withManifest = true;
 		return container;
