+++ packages/astro/src/content/vite-plugin-content-virtual-mod.ts
@@ -3,12 +3,13 @@ import { extname } from 'node:path';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 import glob from 'fast-glob';
 import pLimit from 'p-limit';
-import { type Plugin } from 'vite';
+import type { Plugin } from 'vite';
 import type { AstroSettings } from '../@types/astro.js';
 import { encodeName } from '../core/build/util.js';
 import { AstroError, AstroErrorData } from '../core/errors/index.js';
 import { appendForwardSlash, removeFileExtension } from '../core/path.js';
-import { isServerLikeOutput, rootRelativePath } from '../core/util.js';
+import {  rootRelativePath } from '../core/viteUtils.js';
+import { isServerLikeOutput } from "../core/util.js";
 import type { AstroPluginMetadata } from '../vite-plugin-astro/index.js';
 import {
 	CONTENT_FLAG,
