+++ packages/astro/src/core/util.ts
@@ -1,11 +1,9 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { normalizePath } from 'vite';
 import type { AstroConfig, AstroSettings, RouteType } from '../@types/astro.js';
 import { SUPPORTED_MARKDOWN_FILE_EXTENSIONS } from './constants.js';
-import type { ModuleLoader } from './module-loader/index.js';
-import { prependForwardSlash, removeTrailingForwardSlash, slash } from './path.js';
+import { removeTrailingForwardSlash, slash } from './path.js';
 
 /** Returns true if argument is an object of any prototype/class (but not null). */
 export function isObject(value: unknown): value is Record<string, any> {
@@ -184,52 +182,9 @@ export function relativeToSrcDir(config: AstroConfig, idOrUrl: URL | string) {
 	return id.slice(slash(fileURLToPath(config.srcDir)).length);
 }
 
-export function rootRelativePath(
-	root: URL,
-	idOrUrl: URL | string,
-	shouldPrependForwardSlash = true
-) {
-	let id: string;
-	if (typeof idOrUrl !== 'string') {
-		id = unwrapId(viteID(idOrUrl));
-	} else {
-		id = idOrUrl;
-	}
-	const normalizedRoot = normalizePath(fileURLToPath(root));
-	if (id.startsWith(normalizedRoot)) {
-		id = id.slice(normalizedRoot.length);
-	}
-	return shouldPrependForwardSlash ? prependForwardSlash(id) : id;
-}
 
 export function emoji(char: string, fallback: string) {
 	return process.platform !== 'win32' ? char : fallback;
-}
-
-/**
- * Simulate Vite's resolve and import analysis so we can import the id as an URL
- * through a script tag or a dynamic import as-is.
- */
-// NOTE: `/@id/` should only be used when the id is fully resolved
-export async function resolveIdToUrl(loader: ModuleLoader, id: string, root?: URL) {
-	let resultId = await loader.resolveId(id, undefined);
-	// Try resolve jsx to tsx
-	if (!resultId && id.endsWith('.jsx')) {
-		resultId = await loader.resolveId(id.slice(0, -4), undefined);
-	}
-	if (!resultId) {
-		return VALID_ID_PREFIX + id;
-	}
-	if (path.isAbsolute(resultId)) {
-		const normalizedRoot = root && normalizePath(fileURLToPath(root));
-		// Convert to root-relative path if path is inside root
-		if (normalizedRoot && resultId.startsWith(normalizedRoot)) {
-			return resultId.slice(normalizedRoot.length - 1);
-		} else {
-			return '/@fs' + prependForwardSlash(resultId);
-		}
-	}
-	return VALID_ID_PREFIX + resultId;
 }
 
 export function resolveJsToTs(filePath: string) {
@@ -242,17 +197,7 @@ export function resolveJsToTs(filePath: string) {
 	return filePath;
 }
 
-/**
- * Resolve the hydration paths so that it can be imported in the client
- */
-export function resolvePath(specifier: string, importer: string) {
-	if (specifier.startsWith('.')) {
-		const absoluteSpecifier = path.resolve(path.dirname(importer), specifier);
-		return resolveJsToTs(normalizePath(absoluteSpecifier));
-	} else {
-		return specifier;
-	}
-}
+
 
 /**
  * Set a default NODE_ENV so Vite doesn't set an incorrect default when loading the Astro config
