+++ packages/astro/src/jsx/babel.ts
@@ -2,7 +2,7 @@ import type { PluginObj } from '@babel/core';
 import * as t from '@babel/types';
 import { AstroError } from '../core/errors/errors.js';
 import { AstroErrorData } from '../core/errors/index.js';
-import { resolvePath } from '../core/util.js';
+import { resolvePath } from '../core/viteUtils.js';
 import type { PluginMetadata } from '../vite-plugin-astro/types.js';
 
 const ClientOnlyPlaceholder = 'astro-client-only';
