+++ packages/astro/src/vite-plugin-astro-server/pipeline.ts
@@ -12,18 +12,16 @@ import type {
 } from '../@types/astro.js';
 import { getInfoOutput } from '../cli/info/index.js';
 import { type HeadElements } from '../core/base-pipeline.js';
-import { shouldAppendForwardSlash } from '../core/build/util.js';
 import { ASTRO_VERSION, DEFAULT_404_COMPONENT } from '../core/constants.js';
 import { enhanceViteSSRError } from '../core/errors/dev/index.js';
-import { RewriteEncounteredAnError } from '../core/errors/errors-data.js';
-import { AggregateError, AstroError, CSSError, MarkdownError } from '../core/errors/index.js';
+import { AggregateError,  CSSError, MarkdownError } from '../core/errors/index.js';
 import type { Logger } from '../core/logger/core.js';
 import type { ModuleLoader } from '../core/module-loader/index.js';
-import { prependForwardSlash, removeTrailingForwardSlash } from '../core/path.js';
 import { Pipeline, loadRenderer } from '../core/render/index.js';
-import { DEFAULT_404_ROUTE, default404Page } from '../core/routing/astro-designed-error-pages.js';
+import {  default404Page } from '../core/routing/astro-designed-error-pages.js';
 import { findRouteToRewrite } from '../core/routing/rewrite.js';
-import { isPage, isServerLikeOutput, resolveIdToUrl, viteID } from '../core/util.js';
+import { isPage, isServerLikeOutput, viteID } from '../core/util.js';
+import { resolveIdToUrl } from "../core/viteUtils.js"
 import { PAGE_SCRIPT_ID } from '../vite-plugin-scripts/index.js';
 import { getStylesForURL } from './css.js';
 import { getComponentMetadata } from './metadata.js';
